+++ src/Quartz.Examples.AspNetCore/Startup.cs
@@ -108,7 +108,7 @@ namespace Quartz.Examples.AspNetCore
                 q.ScheduleJob<ExampleJob>(trigger => trigger
                     .WithIdentity("Combined Configuration Trigger")
                     .StartAt(DateBuilder.EvenSecondDate(DateTimeOffset.UtcNow.AddSeconds(7)))
-                    .WithDailyTimeIntervalSchedule(x => x.WithInterval(10, IntervalUnit.Second))
+                    .WithDailyTimeIntervalSchedule(interval: 10, IntervalUnit.Second)
                     .WithDescription("my awesome trigger configured for a job with single call")
                 );
 
@@ -174,7 +174,7 @@ namespace Quartz.Examples.AspNetCore
                     .WithIdentity("Daily Trigger")
                     .ForJob(jobKey)
                     .StartAt(DateBuilder.EvenSecondDate(DateTimeOffset.UtcNow.AddSeconds(5)))
-                    .WithDailyTimeIntervalSchedule(x => x.WithInterval(10, IntervalUnit.Second))
+                    .WithDailyTimeIntervalSchedule(interval: 10, IntervalUnit.Second)
                     .WithDescription("my awesome daily time interval trigger")
                     .ModifiedByCalendar(calendarName)
                 );