+++ src/Quartz.Examples.Worker/Program.cs
@@ -62,7 +62,7 @@ namespace Quartz.Examples.Worker
                         q.ScheduleJob<ExampleJob>(trigger => trigger
                             .WithIdentity("Combined Configuration Trigger")
                             .StartAt(DateBuilder.EvenSecondDate(DateTimeOffset.UtcNow))
-                            .WithDailyTimeIntervalSchedule(x => x.WithInterval(10, IntervalUnit.Second))
+                            .WithDailyTimeIntervalSchedule(interval: 10, intervalUnit: IntervalUnit.Second)
                             .WithDescription("my awesome trigger configured for a job with single call")
                         );